reverted file structure of style.js
