+++ src/App.test.js
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/react";
 import App from "./App";
-import React from "react";
+import React from 'react';
 
 test("renders learn react link", () => {
   render(<App />);
