+++ src/components/wishlistCard/index.js
@@ -7,10 +7,44 @@ import {
   Typography,
 } from "@material-ui/core";
 import React from "react";
-import useStyles from "./style";
 import ButtonComponent from "../Button/ButtonComponent";
 import { DeleteOutline } from "@material-ui/icons";
 import { useHistory } from "react-router";
+
+import { makeStyles } from "@material-ui/core/styles";
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    color: "#fff",
+    margin: 0,
+    // marginTop: '8px',
+    borderRadius: "8px",
+    "&:hover": {
+      cursor: "pointer",
+      transform: "translateY(-2px)",
+    },
+  },
+  cardHeader: {
+    backgroundColor: "#A60000",
+    color: "#fff",
+  },
+  icon: {
+    color: "#fff",
+    padding: "4px",
+    fontSize: "14px",
+  },
+  title: {
+    fontWeight: 700,
+    fontSize: "12px",
+  },
+  subheader: {
+    fontStyle: "italic",
+    fontSize: "10px",
+    color: "rgb(255, 255, 255, 0.8)",
+    marginTop: "8px",
+    marginBottom: "-8px",
+  },
+}));
 
 const WishlistCard = () => {
   const classes = useStyles();
